Constants: rename photoSwipe to photoViewer
