--- src/constants/photoswipe/index.ts
@@ -1,7 +1,0 @@
-export const defaultLivePhotoDefaultOptions = {
-    click: () => {},
-    hide: () => {},
-    show: () => {},
-    loading: false,
-    visible: false,
-};
